Introduce new fee model
The new fee model is similar to eip-1559. It introduces new transaction entities: `MaxPriorityFeePerGas` and `MaxFeePerGas`.
This patch is the initial phase of the migrating to this model. It introduces these two fields for all components of the project
and introduces a new algorithm for base fee calculation.
In golang functions, new fields are grouped into `FeePack` struct which, in turn, is passed to the functions.
This is convenient for further refactoring, as we can only change this struct while keeping the functions signatures unchanged.

Also, `estimateGas` for deployment transactions in niljs have been fixed. Previously, they were passed as execution transactions.