[consensus] run consensus on main shard by all validators
This patch enables consensus for the main shard. Now, it is validated by all available validators.
The list of validators is simply a collection of validators from all shards without duplicates.
The main issue is the tests since the sharded suite now has to wait for the zero shard to start ticking.

Need for #143