[consensus] don't start new round until previous is not completely finished
It could cause a race condition.
Imagine following case our validator started RunSequence but for some
reasons syncer got block before InsertProposal.
In that case we need to interrupt consensus sequence. And we do that
in defer (we call context cancel).
But seems some processes don't happen immediately. And we need to wait
for complete termination of RunSequence. Here we add two lines:
the first one cancels context (it will be called second time in defer but
seems it's safe in Go). And the second one just waits for some result from
RunSequence function. This delay should be quite quick.