[execution] use specific main shard hash from block generation
It's not correct to rely on the latest block.
Main block can be committed earlier than current one that can lead to
the case when we read data from a newer source.
We fixed almost the same issue before for `collectGasPrices` function
because it affected consensus. This part can also affect it in some ways.