[types] exclude bloom from receipt/block hashes
Bloom filters are very complex to prove them using ZK cryptography. So it was
decided to exclude them from block hash calculation. This patch do that
for Receipt and Block types.
For receipts we returns Bloom field on jsonrpc level. DB record doesn't contain
this info anymore.
Block stores Bloom field as a metadata (like a siganture). We store it inside DB
but it doesn't affect a hash itself.

That's purely internal change that affects only hashes. From user POV
nothing was changed.

Closes #52