[execution] validate shardId for outgoing messages
Before this patch it was valid to send messages to non-exising shards.
That caused different errors during getting receipts for messages that
contained such messages.

This patch adds some validation for that. Fix is pretty straightforward
since we don't have "nShards" as a separate option we get an array of
gas prices because amount of elements is exactly shard count.

Closes #85