[syncer] fix prev block hash mismatch error on sync
As in some previous patches we faced with an issue that made an
assumption that we should use last block from some chain. However,
it seems not so and we should exactly specify blockHash for processing.
To avoid such failures in future let's add panic here but only in
debug mode. We had the same panic before but it was pretty easy to trigger
it from third-party side.