[execution] unify config accessor management between gas prices and validators
We decided to change approach to configuration management a bit.
At the moment of block generation (with height=h) we have committed
known block (h-1). This block has a link to mainShardHash that is a base
for config accessor.
Such approach is quite consistent and decrease chances to face with an issue
when we don't have an access to main shard block because it's not replicated yet.