[execution] use deterministic time
We had problems with consensus. That's because the same EVM input
produced different output that led to hash mismatch error.
It was pretty hard to debug this issue. It broke consensus on load gen tests
because uniswap sometimes operates with a time.

This patch removes using of `time.Now()` that breaks an idea of repeatable
blockchain history.

Our block.Timestamp is always zero. So I suggest to use block.Id as a logical time.
It also monotonically increases.