[collate] prevent "Transaction conflict" error because of syncer/consensus race
The main idea is to make validator as an entrypoint for proposal insertion.
We insert proposal in two cases: consensus and block replay.

Sometimes we had an error "Transaction conflict" caused by race between
that two modules. So it seems reasonable to add a lock here to
avoid confusing error messages and execution transaction two times.