Sync Committee: Storage limit error handling (Aggregator)
Previous commit introduces capacity limit for the `BlockStorage`.
This commit contains improvements of `Aggregator` to handle new `ErrCapacityLimitReached` error.

The core changes in behaviour are defined in the `Test_Block_Storage_Capacity_Exceeded` test.
It's expected that, in case if capacity limit is reached, new generated batches are not handled by `Aggregator`:

1. No new tasks are created;
2. Latest fetched values are not updated;