[collate] add wait main shard block logic to block replay
It's possible case that during block replay specific block.mainShardHash
is not available. That shouldn't happen normally since we replay main shard
really fast.
But because of network issues or something else it can be a problem.
So this patch adds simple logic that tries to wait expected block using
subscriptions.