[vm] fix state storage modifying by `SLOAD` opcodes
During execution of `SLOAD` instruction we place read key into state cache.
Then, when we commit a block, we write all elements of state cache to the database.
Hence, even without `SSTORE` instruction, state root of an account can be changed.
For empty account, it leads to impossibility of further deployment on this address.
We fix it by not saving non-exist items into the cache.
It has a drawback in the form of performance, but actually we don't need to store
zero values in the storage, instead we should delete them. It will avoid this drawback.
Here is an issue about this #546

Closes #543