Bundle explorer backend into a single .js file
There was a long-standing problem with how we deliver explorer
backend to prod machines. That's because it used to be delivered in the
source form and we executed "npm" at the spot. This introduces flakiness
and generally makes distributing a single deb package difficult.

This diff uses "rollup" JS module to bundle the explorer backend into a
single file which can be just executed with `node explorer.js`.