[tests] fix race in Test_TaskExecutor_Busy_Handler again
That should be fixed in b0ab720db27ae07b2c9315f2cfe84ed049a47111 but failures in CI
says that it didn't help. Seems gorounine in Never function makes several additional iterations.
This fix a bit strange but here we save problematic variable in a variable so after it it's
safe to modify test suite state.