Bundle docs_ai_backend
It looks like it's not very difficult to bundle all necessary npm
dependencies together with the docs_ai_backend. What is needed is just
to specify `output: standalone` in the next.js config.

I've tried to start the server with `node` from the resulting build
and it worked. However, I can't do any in-depth testing and will wait
for Shukhrat to do that part.

testong

fixed the DNS error and made the app launch using the correct port

fixes