Support config files for sync_committee binaries
Currently we pass some secret data via cmd options which is insecure.
This commit introduces a possibility to read the same options from
config files also. It still keeps all the flags, and in case of two
values passed from both a config and an option, the option value will be
used.