[execution] don't inc ExtSeqno in case if tx is not included to block
The previous patch
(cbed7489a815956487b913a4d6ea96922e3ff28d) fixed the
possibility of adding identical transactions to
multiple blocks. However, it introduced a regression.

Transactions that are not included in a block (and
create so-called temporary receipts) could still
modify the state. This breaks replay, since the
account cannot be brought to the correct state
without that transaction.

This patch fixes the issue by rolling back ExtSeqno
if the transaction is not included in a block.

This mainly affects deploy transactions. For
execution transactions, gas is paid if
`verifyExternal` succeeds, so state changes are valid
in that case.