Allow fetching contract info with "nil contract [address]"
We had a long debugging session yesterday with Oleg, trying
to understand where the archive node's state has diverged
from the validator. It would have been a bit easier to figure
out which contract to blame if we could request the hash and
storage root hash of that contract.

Fortunately, we already have the API for fetching debug information
about smart contracts, which Sync Committee uses. I've just added
an extension to our nil cli.

This will fetch and print contract information for the latest block:

```
nil contract [address]
```

And this will look at a particular block:

```
nil contract --block <block-id-or-tag> [address]
```