[GH-766] RawAPI has been redesigned.
Added the ability to operate any number of different ShardAPIs in any combination. This is necessary for fine-tuning which shards in which modes provide which capabilities via the API.

To achieve the goal of flexible configuration, the mechanism for creating connectors for the shard API was reworked: network and debug local. The construction functions were parameterized by the API types for which connectors need to be created.

In addition, the mechanism for creating node API was reworked to improve convenience and hide details on working with individual shard APIs that consumers do not need to know about.