[vm] drop awaitCall precompile
This feature was experimental and it can be reimplemented in pure EVM.
So we decided to drop it. This patch does that.

In docs we replaced awaitCall with sendRequest. It was too hard to do that
in "sharded-voting-protocol". It requires significant rework on this contracts
and they also seems to be untested. Here we just removed this example.